requirements.txt
numpy>=1.24,<2.0
pandas>=2.1,<2.3
scikit-learn>=1.3,<1.5
lightgbm>=4.1,<4.3
xgboost>=1.7,<2.1
catboost>=1.2,<1.3
matplotlib>=3.8,<3.9
seaborn>=0.13,<0.14
pyarrow>=14.0,<15.0
joblib>=1.3,<1.4
tqdm>=4.66,<4.67
mlflow>=2.9,<3.0
optuna>=3.4,<3.5
pyyaml>=6.0,<7.0
jupyterlab>=4.0,<4.2
notebook>=7.0,<7.2
ruff>=0.1,<0.3
black>=23.9,<24.0
pytest>=7.4,<8.2
